Split multus unit tests into several files
To simplify multus unit tests, split it into several files,
based on testing CNI version.
